Save files by actual name, and the right content to .txt files
The .txt files now appear to work properly. All types of files are named
correctly, though the `binary.jpg` does not work as an image to open.

Perhaps we are handling something somewhere that works for texts as
strings but not in a way that works for images, particularly .jpg.